backend/requirements.txt
fastapi
uvicorn[standard]
ortools
pandas
openpyxl
numpy
requests
python-multipart
# sqlalchemy
motor
pymongo
PyJWT